+++ lib/views/home/about_view.dart
@@ -81,7 +81,7 @@ class AboutView extends StatelessWidget {
                     borderRadius: BorderRadius.circular(20),
                   ),
                   child: const Text(
-                    'Version 6.4',
+                    'Version 6.4.1',
                     style: TextStyle(color: Colors.cyanAccent, fontSize: 14),
                   ),
                 ),